Only show the first available pair
When no names are passed to the `show` command, the app will show the
next optimal pair combination for all team members.  Since we'll have
mostly a few pairs in the team, this is probably easier than having to
list individual team names in the command line.

If we want to make sure only team members that are present are available
(e.g. in a large team with multiple people being absent), then we can
pass individual names to the command line and get only those pairs.